refactor(cli): split component.mjs into lib/ modules + lazy command registry
Split the 1273-line component.mjs into focused modules:
- lib/component-discovery.mjs: discoverComponents, findComponentReadme, findComponentSource, resolveImportPath
- lib/component-loader.mjs: loadDocs, mergeTranslation
- lib/component-format.mjs: formatFull, formatCompact, formatBrief, formatProps, formatBriefAll
- lib/component-legacy.mjs: cleanReadme, extract* (README.md backward compat)
- lib/string-utils.mjs: levenshteinDistance, findClosestComponents
- commands/component/index.mjs: registerComponent + re-exports for backward compat

index.mjs now uses a lazy command registry. Each command is loaded via
dynamic import() with try/catch. If one command fails to load (syntax error,
bad dependency), the other commands still work. The broken command shows a
helpful error message instead of crashing the CLI.

All 1436 tests passing. CLI behavior unchanged.
